pyproject.toml
[project]
name = "klp"
version = "0.1.0"
description = "Kulturelle Landpartie website with event listings, locations, and map"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlmodel>=0.0.22",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
    "jinja2>=3.1.4",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.3",
    "apscheduler>=3.10.4",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "black>=24.8.0",
    "ruff>=0.6.0",
]
